Refactor calculation of CorDapp contents into a separate task.
