+++ cordapp-cpk/src/main/kotlin/net/corda/plugins/cpk/VerifyBundle.kt
@@ -5,12 +5,17 @@ import aQute.bnd.osgi.Jar
 import aQute.bnd.osgi.Verifier
 import org.gradle.api.DefaultTask
 import org.gradle.api.InvalidUserDataException
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.FileCollection
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.tasks.Classpath
 import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.PathSensitive
 import org.gradle.api.tasks.PathSensitivity.RELATIVE
 import org.gradle.api.tasks.TaskAction
+import org.gradle.api.tasks.TaskProvider
 import javax.inject.Inject
 
 @Suppress("UnstableApiUsage")
@@ -23,6 +28,22 @@ open class VerifyBundle @Inject constructor(objects: ObjectFactory) : DefaultTas
     @get:PathSensitive(RELATIVE)
     @get:InputFile
     val bundle: RegularFileProperty = objects.fileProperty()
+
+    private val _classpath: ConfigurableFileCollection = objects.fileCollection()
+    val classpath: FileCollection
+        @Classpath
+        @InputFiles
+        get() = _classpath
+
+    /**
+     * Don't eagerly configure the [DependencyCalculator] task, even if
+     * someone eagerly configures this [VerifyBundle] by accident.
+     */
+    internal fun setDependenciesFrom(task: TaskProvider<DependencyCalculator>) {
+        _classpath.setFrom(task.flatMap(DependencyCalculator::externalJars), task.flatMap(DependencyCalculator::dependencies))
+        _classpath.disallowChanges()
+        dependsOn(task)
+    }
 
     @TaskAction
     fun verify() {
