+++ gradle.properties
@@ -20,3 +20,5 @@ bnd_version=5.2.0
 bintray_version=1.8.5
 artifactory_version=4.16.1
 gradle_publish_version=0.12.0
+
+artifactory_contextUrl=https://software.r3.com/artifactory
